fix(client): handle jsonrpc improper redirects
