+++ client/src/utils/helpers/EnhancedState.ts
@@ -37,8 +37,8 @@ import { displaySlice } from "../../features/display/displaySlice";
 import { inactiveKgProjectsApi } from "../../features/inactiveKgProjects/InactiveKgProjectsApi";
 import { kgInactiveProjectsSlice } from "../../features/inactiveKgProjects/inactiveKgProjectsSlice";
 import { kgSearchApi } from "../../features/kgSearch";
-import { datasetFormSlice } from "../../features/project/dataset";
 import projectCloudStorageApi from "../../features/project/components/cloudStorage/projectCloudStorage.api";
+import { datasetFormSlice } from "../../features/project/dataset";
 import { projectCoreApi } from "../../features/project/projectCoreApi";
 import projectGitLabApi from "../../features/project/projectGitLab.api";
 import { projectKgApi } from "../../features/project/projectKg.api";
@@ -47,7 +47,7 @@ import { projectV2Api } from "../../features/projectsV2/api/projectV2.enhanced-a
 import { projectV2NewSlice } from "../../features/projectsV2/new/projectV2New.slice";
 import { recentUserActivityApi } from "../../features/recentUserActivity/RecentUserActivityApi";
 import sessionsApi from "../../features/session/sessions.api";
-import { sessionSidecarApi } from "../../features/session/sidecarApi";
+import sessionSidecarApi from "../../features/session/sidecar.api";
 import startSessionSlice from "../../features/session/startSession.slice";
 import { startSessionOptionsSlice } from "../../features/session/startSessionOptionsSlice";
 import termsApi from "../../features/terms/terms.api";
